+++ weldx/tests/transformations/test_cs_manager.py
@@ -2230,14 +2230,16 @@ def test_assign_data_exceptions(arguments, exception_type, test_name):
         csm.assign_data(*arguments)
 
 
-def test_remove_data():
-    """Test remove data."""
+def test_delete_data():
+    """Test delete data."""
     csm = setup_csm_test_assign_data()
     data_name = "foo"
     csm.assign_data([[1, 2, 3], [3, 2, 1]], data_name, "root")
     assert data_name in csm.data_names
-    csm.remove_data(data_name)
+    csm.delete_data(data_name)
     assert data_name not in csm.data_names
+
+    # TODO: add test case for delete_data and subsystems.
 
 
 # test_has_data_exceptions -------------------------------------------------------------